tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/removing_animation.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - resetting animation-name of running animation</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="Once an animation has started it continues
   until it ends or the 'animation-name' is removed">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            left: 0;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: slide 2s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         setTimeout(function() {
            var passed = getComputedStyle(div, null).left != 'auto';
            div.className = '';
            var passed = passed && getComputedStyle(div, null).left == 'auto';
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         }, 500);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Below there should be a blue square moving right briefly and then
   jumping back to its initial position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
